requirements.txt
aiohttp==3.6.2
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.6
anaconda-project==0.8.2
appdirs==1.4.4
apriori==1.0.0
apyori==1.1.2
asn1crypto==0.24.0
astroid==2.1.0
astropy==3.1
async-timeout==3.0.1
atomicwrites==1.2.1
attrs==21.2.0
Automat==20.2.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.6
beautifulsoup4==4.10.0
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==3.0.2
bokeh==1.0.2
boto==2.49.0
boto3==1.17.7
botocore==1.20.7
Bottleneck==1.2.1
bz2file==0.98
certifi==2021.10.8
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
clyent==1.2.2
colorama==0.4.1
community==1.0.0b1
comtypes==1.1.7
conda==4.10.3
conda-build==3.17.6
conda-package-handling==1.3.11
conda-verify==3.1.1
constantly==15.1.0
contextlib2==0.5.5
crawler==0.0.2
cryptography==2.4.2
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.23
cytoolz==0.9.0.1
dask==1.0.0
datashape==0.5.4
decorator==4.3.0
defusedxml==0.5.0
distributed==1.25.1
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
feedfinder2==0.0.4
feedparser==6.0.1
filelock==3.0.10
Flask==1.0.2
Flask-Cors==3.0.7
Flask-SQLAlchemy==2.4.4
future==0.17.1
gensim==4.1.2
gevent==1.3.7
glob2==0.6
greenlet==0.4.15
h5py==2.8.0
heapdict==1.0.0
html5lib==1.0.1
hyperlink==21.0.0
idna==2.8
imageio==2.4.1
imagesize==1.1.0
importlib-metadata==0.6
incremental==17.5.0
input-mocker==0.1.4
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itemadapter==0.1.0
itemloaders==1.0.3
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.2
jellyfish==0.8.8
jieba3k==0.35.1
Jinja2==2.10
jmespath==0.10.0
joblib==1.0.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.3
jupyterlab-launcher==0.13.1
jupyterlab-server==0.2.0
keras==2.6.0
keyring==17.0.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
libarchive-c==2.8
llvmlite==0.26.0
locket==0.2.0
lxml==4.2.5
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.2
more-itertools==4.3.0
mpmath==1.1.0
msgpack==0.5.6
mtcnn==0.1.1
multidict==4.7.6
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
newspaper3k==0.2.8
nltk==3.6.4
nose==1.3.7
notebook==5.7.4
numba==0.41.0
numexpr==2.6.8
numpy==1.21.2
numpydoc==0.8.0
odo==0.5.1
olefile==0.46
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
openpyxl==2.5.12
packaging==18.0
pandas==0.23.4
pandocfilters==1.4.2
parsel==1.6.0
parso==0.3.1
partd==0.3.9
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pep8==1.7.1
pickleshare==0.7.5
Pillow==5.3.0
pkginfo==1.4.2
pluggy==0.8.0
ply==3.11
prometheus-client==0.5.0
prompt-toolkit==2.0.7
Protego==0.1.16
psutil==5.4.8
py==1.7.0
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
PyDispatcher==2.0.5
pyflakes==2.0.0
Pygments==2.3.1
PyHamcrest==2.0.2
pylint==2.2.2
PyMySQL==0.10.1
pyodbc==4.0.25
pyOpenSSL==18.0.0
pyparsing==2.3.0
PySocks==1.6.8
pyspark==3.1.1
pytest==4.0.2
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.2.0
pytest-openfiles==0.3.1
pytest-remotedata==0.3.1
python-crfsuite==0.9.7
python-dateutil==2.7.5
python-louvain==0.15
pytz==2018.7
pyvi==0.1.1
PyWavelets==1.0.1
pywin32==223
pywinpty==0.5.5
PyYAML==3.13
pyzmq==17.1.2
QtAwesome==0.5.3
qtconsole==4.4.3
QtPy==1.5.2
queuelib==1.5.0
rake-nltk==1.0.6
regex==2021.9.30
requests==2.21.0
requests-file==1.5.1
rope==0.11.0
ruamel-yaml==0.15.46
s3transfer==0.3.6
scikit-image==0.14.1
scikit-learn==0.20.1
scipy==1.1.0
Scrapy==2.3.0
seaborn==0.9.0
segtok==1.5.10
Send2Trash==1.5.0
service-identity==18.1.0
sgmllib3k==1.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
sklearn-crfsuite==0.3.6
smart-open==5.2.1
snowballstemmer==1.2.1
sortedcollections==1.0.1
sortedcontainers==2.1.0
soupsieve==2.2.1
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
spyder==3.3.2
spyder-kernels==0.3.0
SQLAlchemy==1.2.15
statistics==1.0.3.5
statsmodels==0.9.0
sympy==1.3
tables==3.4.4
tabulate==0.8.9
tblib==1.3.2
terminado==0.8.1
testpath==0.4.2
timeout-decorator==0.5.0
tinysegmenter==0.3
tldextract==3.0.0
toolz==0.9.0
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
Twisted==20.3.0
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.24.1
w3lib==1.22.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xlrd==1.2.0
XlsxWriter==1.1.2
xlwings==0.15.1
xlwt==1.3.0
yake==0.4.8
yarl==1.6.0
zict==0.1.3
zope.interface==5.1.0
